print: fix nesting of applications
The issue is obvious when a comment preceeds and attaches to an
application, for instance:

    (* some long comment on its own line *)
    some_fun applied

In this case the comment was inserted under the nest node (it must
remain at the same indentation level as what it's attached to), causing
the whole application to be indented, not just the arguments (which is
what we'd want), e.g.

    (* some long comment on its own line *)
      some_fun applied

This is bad!

Also, the comment ended up at a different indentation level as the
application: this is bad, it might get attached differently next time
around.