normalize: force pipe on first match/function case
N.B. the change is heavy handed because the pipe are not attached to the
case but to its parent.

This is different from how constructor declarations are handled for
instance.

The normalization code, as well as the printing code, would be much
nicer/simpler if pipes were always handled as they are for constructor
decls.