comments: more regular handling of explicitely inserted comments (WIP)
Also stop attaching docstrings as attribute. Just attach the string and
mark the underlying comment as explicitely inserted into the document.

There is one bug: empty docstring disappear, but they are actually
needed in some case (to force another docstring to attach to a parent
for instance).
In those case the printed code parses back to a different ast.